Renaming "basic_resolution" to "pose_stack_construction" for clarity
