+++ tmol/__init__.py
@@ -4,7 +4,7 @@ from tmol.io.canonical_ordering import (  # noqa: F401
     ordered_canonical_aa_atoms_v2,
     max_n_canonical_atoms,
 )
-from tmol.io.basic_resolution import pose_stack_from_canonical_form  # noqa: F401
+from tmol.io.pose_stack_construction import pose_stack_from_canonical_form  # noqa: F401
 from tmol.score import beta2016_score_function  # noqa: F401
